Implement Smoothie spec for testing RSpec output
Pages 30-31

This spec has one passing and one failing example. In this case,
`--next-failure` flag and `--only-failures` flags will run similarly,
but if there were multiple failures, `--next-failure` would only output
one (the next one).

I used this command [1] to save the html to a file that I could then
open in a browser window (by clicking on the file in finder) to see the
output.

RSpec's help flag [2] is really useful for showing a list of built-in
output formats.

[1] rspec --format html --out ./reports/rspec_results.html
[2] rspec --help