Chapter 2 (#4)
* Create coffee spec

Pages 15-19

I decided to actually make a tea spec, but didn't rename the file... You
can decide if this was intentional to attempted to line up with the book,
or not 😅

* Example of slow specs

Page 20

Can use the profile flag to return the slowest specs. By default profile
will return the 10 slowest, but you can pass in an option number.

Example usage:
rspec --profile
rspec --profile 3
rspec --p
rspec --p 2

* Add RSpec configuration to enable runtime options

Pages 23-24

On page 23 we're specifically looking at using the --only-failures flag.

Once adding the configuration, which would typically live in its own
file, run the specs normally so the passing/failing status is recorded.
From there, us use the --only-failures flag.

Example usage:
rspec --only-failures

* Add oat_milk logic to price to method

Page 24

Note that this is my implementation, not what is in the
book on page 24 🙂

* Introduce focusing to RSpec configuration, and use it

Pages 24-26

Adding an `f` to the beginning of a the Rspec method name will focus on
that specific test [1]. This is useful when you are working
on a specific test and want to run only that test.

[1] `context` becomes `fcontext`, `it` becomes `fit`, etc

* Remove focus from context

Page 26

* Use RSpec's pending method for marking WIP specs

Pages 26-29

Everything below pending will be run by RSpec, but won't result in a
failure if the test doesn't pass.

You can also write out just an `it` statement [1] (without the block)
to  sketch out incoming ogic.

[1] `it 'is lighter in colour'`

* Add logic so spec should pass, denoted with a failure message

Page 29

* Implement Smoothie spec for testing RSpec output

Pages 30-31

This spec has one passing and one failing example. In this case,
`--next-failure` flag and `--only-failures` flags will run similarly,
but if there were multiple failures, `--next-failure` would only output
one (the next one).

I used this command [1] to save the html to a file that I could then
open in a browser window (by clicking on the file in finder) to see the
output.

RSpec's help flag [2] is really useful for showing a list of built-in
output formats.

[1] rspec --format html --out ./reports/rspec_results.html
[2] rspec --help