Improve setup of api_spec
Pages 67-69

We'll eventually move the RecordResult class into the application code,
but for now it can live in the spec.

Using a simple value object like a RecordResult or Struct between layers
makes it easier to isolate code and trust your tests. See Gary
Bernhardt's 'Boundaries' talk [1] for more details.

[1] https://www.destroyallsoftware.com/talks/boundaries