Chapter 5 (#6)
* Create chapter 5 folder with initial code

Most of which is a copy from how chapter 4 ended (minus the extra work
I did to make the get request tests pass).

* Sketch out first unit test for API

Pages 63 - 66

Used dependency injection (DI) to create a Ledger instance; we'll
eventually use the Ledger to store data, but at the moment we're only
concerned about testing the api.

* Improve setup of api_spec

Pages 67-69

We'll eventually move the RecordResult class into the application code,
but for now it can live in the spec.

Using a simple value object like a RecordResult or Struct between layers
makes it easier to isolate code and trust your tests. See Gary
Bernhardt's 'Boundaries' talk [1] for more details.

[1] https://www.destroyallsoftware.com/talks/boundaries

* Add spec for receiving a 200 response

PAges 69 - 70

* Refactor specs in first context

Page 72

From the book:
It may be tempting to DRY them up even further by also mocing the
`post '/expenses...` lines into the `before` hook. Such a move would be
overkill, though.

Putting the `post` in the `before` hook will likely get in the way of
future specs for this context. For instance, adding support for XML
clients would require a different HTTP header. If you send the HTTP
request from your setup code, you won't be able to twaek the headers for
just your XML examples.

* Add specs and code to handle failed post requests

Pages 73-72

* Implement Ledger class

Pages 75-77

* Refactor parsing JSON into a let block

Exercise 1, page 78

I first did this with a helper method, but I prefer using a let block
when the logic isn't too complicated.

* Add spec outline for GET request

page 78

* Implement specs for GET request

pages 78-79

I don't love my use of the eq matcher, but I think due to time
constraints, and knowing that we'll be covering the book's
implementation shortly, I'm going to leave it as is.