Chapter 6 (#7)
* Setup code for chapter 6

Mostly copy and pasted from the chapter 5 code, with the exception of
replacing webrick with puma.

* Hook up the database

pages 82-83

This app will use a SQLite database to store expenses. The database will
be managed by Sequel, a Ruby ORM.

* Enable creation of separate environment databases

pages 83-84

I ran [1] in the command line to run the migration, which created the
development database.

[1] bundle exec sequel -m ./db/migrations sqlite://db/developement.db --echo

* Implement suite level hook for database setup

page 85

We've put this 'partially shared' code in spec/support.db so we can
'require' it when needed.

This means our unit specs, which don't need to have any database setup,
will stay nice and fast.

* Implement first test for Ledger

pages 86-87

* Use aggregate_failures to display all failures in a single run

pages 87-88

* Implement record method in Ledger class

pages 88-89

This gets the first spec in the ledger_spec.rb file to pass. 🎉

* Introduce spec testing an invalid expense

pages 89-92

* Isolate specs using database transactions

pages 92-94

* Implement invalid expense logic

pages 94-95

* Implement specs and logive for expenses_on method

pages 95-97

* Remove pending from expense_tracker_api_spec.rb

page 98

Now that we've implemented the logic in the Ledger class, we can remove
the pending from the spec.

* Get the application to work for reals

pages 98-101

Changes I needed to make, outside of the book's instructions of requiring
sequel.rb in ledger.rb:

- re-run the migration file, this time spelling development correctly 🤦‍♀️ [1]
- use a different curl command [2] to make a POST request, rather than
the book's [3]. Pretty confident this has everything to do with using
Puma, while the book is using Webrick
- get rid of `body.read.rewind`, which I needed when I was using Webrick

The GET request from the book [4] worked without issue.

[1] be sequel -m ./db/migrations sqlite://db/development.db --echo
[2] curl -X POST localhost:9292/expenses -H "Content-Type: application/json" --data '{"payee":"Zoo", "amount":10, "date":"2017-06-10"}' -w "\n"
[3] curl localhost:9292/expenses --data '{"payee":"Zoo", "amount":10, "date":"2017-06-10"}' -w "\n"
[4] curl localhost:9292/expenses/2017-06-10 -w "\n"

* Add rewind back in so the specs pass...

* Improve validation of posting expenses

pages 101 - 102