Use configure.include to include in all specs
page 121

If you need to include the same helper module into many or all of your
specs, you can call configure.include in an RSpec.configure block.

This example will include the APIHelpers module in EVERY example group,
which we likely don’t want, but it does means we can get rid of the
include we had in the previously commit.

The book tells us we’ll go over how to share code conditionally on page
140, so something to look forward to.