Chapter 7 (#8)
* Setup code for chapter 7

* Create alias for using binding.pry

pages 111-112

Adding 'p' in front of an it or describe block will result in
binding.pry being run at the end of the block.

* Extract 'glue code' from the acceptance test into a helper module

page 121

This can then be included in the acceptance tests that require it.

* Use configure.include to include in all specs

page 121

If you need to include the same helper module into many or all of your
specs, you can call configure.include in an RSpec.configure block.

This example will include the APIHelpers module in EVERY example group,
which we likely don’t want, but it does means we can get rid of the
include we had in the previously commit.

The book tells us we’ll go over how to share code conditionally on page
140, so something to look forward to.

* Change APIHelpers from module to shared_context

pages 122-124

This allows us to include a before hook which uses the basic_authorize
method from Rack::Test to treat the HTTP request as coming from a
logged-in user.

For now, I'm continuing to use RSpec.configure to include the shared
context in the all specs.

* Add addressable gem to Gemfile

pages 128-131

Note that you need to add the 'addressable' gem to the Gemfile (and
then run 'bundle install') to run the tests in addressable_spec.rb.

* Refactor shared examples into shared_examples.rb

pages 128-131