Chapter 13 (#15)
* Add starting code for the exercises

* Create an RSpec block that would check constant names...

using the verify doubled contestants method, but only when you want it
to using a flag [1].

[1] rspec -rtrail_map -rsupport/verify_doubled_constants

* Use the class name, rather than a string

My preference is to use the class name, rather than a string, when
creating a new instance of a class. This is because it makes the code
easier to read and understand, and it also helps to avoid potential
errors that can occur if the class name is changed in the future.